# specs/001-haruhanjang-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 하루한장 API
  version: 0.1.0
  description: |
    사장님 PWA(S1~S8)와 기관 대시보드(S7)가 쓰는 REST API.
    근거: spec.md FR-xxx, design/SD_01 IPO, SD_02 화면, SD_03 테이블·뷰.

    공통 규칙
    - 모든 경로는 `/api` 아래. 인증은 HttpOnly 쿠키(`hrh_session`, 서명 JWT).
    - 날짜는 KST 기준 `YYYY-MM-DD`.
    - 게이트 차단은 HTTP 409 + `GateBlocked` 본문. 화면은 `gate` 코드로 SD_02 §11-2 표준 문구를 고른다.
    - 사장님 API는 로그인한 계정의 가게 데이터만 다룬다(FR-091). 경로에 store_id를 받지 않는다.
    - 기관 API(`/api/org/*`)는 익명 집계 뷰만 조회한다(FR-076, research R7).
    - 인가(RBAC): 각 경로의 권한은 data-model §7-2. 권한이 없으면 403 Forbidden, 게이트 차단은 409.
servers:
  - url: /api
...
paths:
...
  /auth/kakao/callback:
    get:
      tags: [auth]
      summary: 카카오 콜백. 기존 계정이면 로그인, 없으면 가입 대기 세션 발급
      parameters:
        - { in: query, name: code, required: true, schema: { type: string } }
        - { in: query, name: state, required: true, schema: { type: string } }
      responses:
        '302': { description: '가입 대기면 /start(S1), 가입 완료면 /today(S2)' }
  /auth/dev-login:
    post:
      tags: [auth]
      summary: 개발 전용 로그인 (AUTH_DEV_LOGIN=1 일 때만 존재)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [devSubject], properties: { devSubject: { type: string } } }
      responses:
        '204': { description: '세션 발급' }
        '404': { description: '운영 환경에서는 없음' }
...
  /org/auth/login:
    post:
      tags: [org]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [loginId, password], properties: { loginId: { type: string }, password: { type: string } } }
      responses:
        '204': { description: '기관 세션 발급' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /staff/auth/login:
    post:
      tags: [staff]
      summary: 운영 인력 로그인 (아이디별 5회 실패 시 15분 잠금)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [loginId, password], properties: { loginId: { type: string }, password: { type: string } } }
      responses:
        '204': { description: 'staff 세션 발급' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { description: '잠금 중' }
...
components:
  responses:
    Unauthorized:
      description: 로그인 필요
...